% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dmnormMax.R
\name{DMnorm_max}
\alias{DMnorm_max}
\title{Joint multivariate normal probability (max)}
\usage{
DMnorm_max(x, m, C, K)
}
\arguments{
\item{x}{-- data given in a matrix format, where rows are n cells, and columns are q genes}

\item{m}{-- mean vector}

\item{C}{-- covariance matrix q x q}

\item{K}{-- number of clusters}
}
\value{
Returns a joint probability of data (i.e. sum of log-density values)
}
\description{
Takes in n data points and outputs their joint probability under multivariate normal distribution. This function is used by MergeM function.
}

