% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeMax.R
\name{MergeM}
\alias{MergeM}
\title{Function for merging two clusters (max)}
\usage{
MergeM(dat, cl_id, cent, K)
}
\arguments{
\item{dat}{-- data given as matrix, where in rows are n cells, and in columns are q genes}

\item{cl_id}{-- cluster indicator variable}

\item{cent}{-- a matrix, where each row is centre/mean of each cluster}

\item{K}{-- number of clusters}
}
\value{
Returns indeces of clusters that should be merged
}
\description{
Takes in K number of clusters and outputs the indeces of clusters that will be merged. Clusters for merging are selected based on largest probability.
}

