% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca_reduce.R
\name{PCAreduce}
\alias{PCAreduce}
\title{Function to perform clustering and dimensionality reduction}
\usage{
PCAreduce(D_t, nbt, q, method)
}
\arguments{
\item{D_t}{-- dataset, presented as marix, where rows are cells, and colms are genes}

\item{nbt}{-- number of samples, i.e. number of times to repeat pcaReduce framework}

\item{q}{-- number of dimensions to start with}

\item{method}{-- can be S or M, S - to perform sampling based merging, M - to perform merging based on largest probability}
}
\value{
Returns a list of length nbt; each item in the list is a matrix containing cell allocation variables
}
\description{
Takes in your dataset (a matrix with rows corresponding to cells, and columns corresponding to genes), performs prespecified dimensionality reduction and merning.
}

